// packages/ui/src/types/sizes.ts
export type NormalSizes = "xs" | "sm" | "md" | "lg" | "xl";
export type ExtraSizes =
  | "2xl"
  | "3xl"
  | "4xl"
  | "5xl"
  | "6xl"
  | "7xl"
  | "8xl"
  | "9xl";
export type TextSizes = NormalSizes | ExtraSizes;

export type RoundedSizes = "lg" | "xl" | "full";

export type TextWeights =
  | "thin"
  | "extralight"
  | "light"
  | "normal"
  | "medium"
  | "semibold"
  | "bold"
  | "extrabold"
  | "black";

export type Spans =
  | "auto"
  | 1
  | 2
  | 3
  | 4
  | 5
  | 6
  | 7
  | 8
  | 9
  | 10
  | 11
  | 12
  | "full";

export type Gap = 0 | 1 | 2 | 3 | 4;

export type Spaces =
  | 0
  | 0.5
  | 1
  | 1.5
  | 2
  | 2.5
  | 3
  | 4
  | 5
  | 6
  | 7
  | 8
  | 9
  | 10
  | 11
  | 12
  | 14
  | 16
  | 18
  | 20;

export type Leadings =
  | 3
  | 4
  | 5
  | 6
  | 7
  | 8
  | 9
  | 10
  | "none"
  | "tight"
  | "snug"
  | "normal"
  | "relaxed"
  | "loose";